# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory System API",
        "description": "Documentación de la API del sistema de inventarios",
        "contact": {
            "email": "dev@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/bodegas": {
            "get": {
                "tags": [
                    "Bodegas"
                ],
                "summary": "Obtener listado de todas las bodegas",
                "description": "Retorna todas las bodegas con su responsable, ordenadas por nombre",
                "operationId": "d9fbea0fa476c4bf29c890a2e246944e",
                "responses": {
                    "200": {
                        "description": "Listado de bodegas exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Bodega Principal"
                                                    },
                                                    "id_responsable": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "estado": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_by": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "responsable": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Responsable Bodega"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "bodega1@inventory.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bodegas listadas exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al listar bodegas: Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Bodegas"
                ],
                "summary": "Crear una nueva bodega",
                "description": "Crea una nueva bodega en el sistema",
                "operationId": "5847d772da639981faf9de009bb02961",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "id_responsable"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre de la bodega (máximo 30 caracteres)",
                                        "type": "string",
                                        "maxLength": 30,
                                        "example": "Bodega Norte"
                                    },
                                    "id_responsable": {
                                        "description": "ID del usuario responsable (debe existir en la tabla users)",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "created_by": {
                                        "description": "ID del usuario que crea el registro (default: 1)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "estado": {
                                        "description": "Estado de la bodega (default: 1 = activo)",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bodega creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Bodega Norte"
                                                },
                                                "id_responsable": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "estado": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_by": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "responsable": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Responsable Bodega"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "bodega1@inventory.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "creador": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Admin Principal"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "admin@inventory.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bodega creada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "properties": {
                                                "nombre": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo nombre es obligatorio."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear bodega: Mensaje de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}